[refactor] home 페이지 index로 변경
- home 페이지를 index 페이지로 변경
