+++ fe/src/components/organisms/SideBar/index.tsx
@@ -24,8 +24,8 @@ interface ButtonProps {
 }
 
 const TITLE: Array<ButtonProps> = [
-  { id: 0, text: '', icon: Twitter({ width: '40px', height: '40px' }), link: '/home' },
-  { id: 1, text: '홈', icon: Home({ width: '30px', height: '30px' }), link: '/home' },
+  { id: 0, text: '', icon: Twitter({ width: '40px', height: '40px' }), link: '/' },
+  { id: 1, text: '홈', icon: Home({ width: '30px', height: '30px' }), link: '/' },
   {
     id: 2,
     text: '탐색하기',
